+++ packages/core/src/router/default_routes.ts
@@ -1,8 +1,7 @@
 import { createEndpoint } from "@colyseus/better-call";
 import { createRouter } from "./index.ts";
 import * as matchMaker from "../MatchMaker.ts";
-import { getBearerToken } from "../utils/Utils.ts";
-import { getTransport } from "../Transport.ts";
+import { createAuthContext, getTransport } from "../Transport.ts";
 
 export const postMatchmakeMethod = createEndpoint("/matchmake/:method/:roomName", { method: "POST" }, async (ctx) => {
   // do not accept matchmaking requests if already shutting down
@@ -21,12 +20,7 @@ export const postMatchmakeMethod = createEndpoint("/matchmake/:method/:roomName"
       method,
       roomName,
       clientOptions,
-      {
-        token: getBearerToken(ctx.request.headers.get('authorization')),
-        headers: ctx.request.headers,
-        ip: requestHeaders.get('x-forwarded-for') ?? requestHeaders.get('x-client-ip') ?? requestHeaders.get('x-real-ip'),
-        req: ctx.request as any,
-      },
+      createAuthContext({ headers: requestHeaders, req: ctx.request }),
     );
 
     //